sdv_amazon_stream/requirements.txt
pandas
sdv
kafka-python